+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.30.0] - 2023-08-28
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Intro state.
+- [@mjkuranda](https://github.com/mjkuranda): Logo image.
+- [@mjkuranda](https://github.com/mjkuranda): Credits for fonts.
 
 ## [0.29.0] - 2023-08-21
 ### Changed